Fixes some issues in asset loading and a int conversion error
Fixes a conversion error in the StackAllocator.

Makes the OpenGL function lookup more robust.

+++ src/MemoryAllocators/StackAllocator.cpp
@@ -47,6 +47,6 @@ namespace Capstan
             assert(false);
         }
 
-        this->marker = (Byte *)allocation.memory - (Byte *)this->start;
+        this->marker = (UInt32)((Byte *)allocation.memory - (Byte *)this->start);
     }
 }
